requirements/requirements.txt
pip
wheel
amqp
beautifulsoup4
bs4
billiard
celery
chardet
cryptography
dateparser
elasticsearch
feedparser
html5lib
openpyxl
Pillow==4.1.1
pyelasticsearch
pyexcel==0.5.2
pyexcel-io==0.4.3
requests==2.14.2
SQLAlchemy==1.1.9
xlrd==1.0.0
XlsxWriter==0.9.6
python-dotenv
